requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
bitsandbytes==0.*
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
datasets==3.1.0
diffusers==0.31.0
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
jaxlib==0.4.13
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
pathspec==0.12.1
peft==0.13.2
pillow==10.4.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==3.20.3
psutil==6.1.0
pyarrow==17.0.0
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
ruff==0.7.3
safetensors==0.4.5
scipy==1.10.1
sentry-sdk==2.18.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sounddevice==0.5.1
sympy==1.13.3
termcolor==2.4.0
tokenizers==0.20.3
tomli==2.1.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.0
transformers==4.46.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.7
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
